PlanRepository.java edited online with Bitbucket

+++ school-commons/src/main/java/pl/sda/reposiotory/PlanRepository.java
@@ -3,5 +3,5 @@ package pl.sda.reposiotory;
 import org.springframework.data.repository.CrudRepository;
 import pl.sda.model.Plan;
 
-public interface PlanRepository extends CrudRepository<Plan, Long> {
+public interface PlanRepository extends CrudRepository<Plan, Integer> {
 }
